Update README.md
A deep learning project that classifies bone X-ray images to detect fractures. The model is trained to distinguish between fractured and non-fractured bones and is deployed using a Streamlit application for an interactive user experience. This project is focused on building and deploying a deep learning model that can classify bone X-ray images as either containing a fracture or not. The model is implemented using TensorFlow and Keras, and the web interface is built with Streamlit.